made script specific to ci and changed to only store data about the module being deployed
